Gun class Input Section Fix
기존에는 총이 독자적으로 IMC를 들고 있었는데 해당 구조가 적까지 총을 사용하는 것까지 고려하면 부적격하다고 판단. 이후 총은 Input을 받지않고 작동 가능한 함수만 들고 있는 상태에서 다른 Character들이 호출하는 식으로 변경. 구조는 GPT가 추천해줌

+++ Source/StealthFPS/DataAssets/GunDataAsset.h
@@ -28,7 +28,7 @@ public:
 	float Distance;
 	
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Ammo)
-	uint8 CurrentAmmo;
+	int32 CurrentAmmo;
 	
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Ammo)
 	uint8 MaxAmmo;
@@ -37,7 +37,7 @@ public:
 	uint8 SubAmmo;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Ammo)
-	uint8 SubMaxAmmo;
+	int32 SubMaxAmmo;
 
 	UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = Ammo)
 	float FireSpeed;
